refactor(adr-061): address review findings on binary_naming/stable_abi/comparability
Three Codex findings on the previous commit, all valid:

1. abicheck/binary_naming.py was brand-new code registered as a legacy
   flat-root module, which lets the architecture gate treat it as
   grandfathered migration debt it isn't. Unlike qualified_name_segments.py
   (which predates ADR-061 classification and has real existing callers
   depending on its flat import path), this file has no pre-existing root
   import path to stay compatible with -- move it to
   abicheck/model/binary_naming.py directly and update every importer
   (binary_utils.py's back-compat re-export, the five diff_*.py callers,
   and its own test) to the real path.

2. stable_abi.py is not solely a shared data table: classify() evaluates
   imports against the stable-ABI floor to produce findings, and
   parse_abi3_version() is frontend-input parsing. Labeling the whole
   mixed module model would misrepresent those higher-layer algorithms as
   valid innermost dependencies. Reverted to unclassified pending a real
   split into its data half (model) and algorithm half (compare/frontends).

3. comparability_fields.py/comparability_json.py/comparability_language_mode.py/
   comparability_sequences.py are comparability.py's own private helper
   family, and comparability.py itself stays unclassified because its
   role straddles compare (field-matching/carve-out logic) and policy
   (the gate itself). Locking the helpers into policy now would create a
   forbidden compare -> policy edge the moment comparability.py is
   eventually classified compare. Reverted all four to unclassified, to
   be decided together with comparability.py as one family.

Updated both changelog fragments (including the earlier, already-pushed
one) to reflect the corrected final state.

Verification: check_architecture.py 0 errors; mypy/ruff clean; targeted
tests (632) plus the architecture suite (114) all pass.
