+++ abicheck/diff_cpp_patterns.py
@@ -48,7 +48,7 @@ from collections.abc import Iterable, Mapping
 from dataclasses import dataclass
 from typing import TYPE_CHECKING
 
-from .binary_naming import strip_vendor_hash
+from .model.binary_naming import strip_vendor_hash
 from .checker_policy import ChangeKind, ReachabilityState
 from .checker_types import Change
 from .diff_helpers import make_change
