+++ abicheck/diff_versioning.py
@@ -20,7 +20,7 @@ diffing, SONAME bump recommendations, and version-script-missing detection.
 
 from __future__ import annotations
 
-from .binary_naming import strip_vendor_hash
+from .model.binary_naming import strip_vendor_hash
 from .checker_policy import API_BREAK_KINDS, BREAKING_KINDS, ChangeKind, Verdict
 from .checker_types import Change
 from .diff_helpers import make_change
