+++ abicheck/diff_wheel_deployment.py
@@ -39,7 +39,7 @@ system libraries this module doesn't enumerate — see "Out of scope".
 
 from __future__ import annotations
 
-from .binary_naming import strip_vendor_hash
+from .model.binary_naming import strip_vendor_hash
 from .checker_policy import ChangeKind
 from .checker_types import Change
 from .diff_helpers import make_change
