+++ changelog.d/1788015192_claude_adr061_remaining_classification.md
@@ -1,3 +1,3 @@
 ### Changed
 
-- **17 more flat root modules given their ADR-061 responsibility-package classification** — `architecture/modules.yaml`'s `legacy_paths` previously left 57 flat root modules with no `may_import`-direction enforcement at all (a module not physically inside its target package *and* absent from every layer's `legacy_paths` is invisible to `check_architecture.py`'s dependency-direction check, per `_source_layer_for`). This closes a verified-safe subset of that gap, greedily selected so the resulting graph introduces zero new `dependency-direction`/`dependency-cycle`/`unclassified-import` findings (checked against the real repo-wide import graph, not a manual read): `model` gains `qualified_name_segments.py` (a dependency-free leaf shared by `compare`'s `diff_namespaces.py` and `extract`'s `dumper.py`-family callers, reclassified out of `compare` since `extract` may not import `compare`) and `stable_abi.py`; `extract` gains `dumper_manifest.py`; `policy` gains `comparability_fields.py`/`comparability_json.py`/`comparability_language_mode.py`/`comparability_sequences.py`, `pattern_verdicts.py`, and the three root-level back-compat shims (`contract_coverage_exit.py`, `exit_decision.py`, `severity.py`) whose own docstrings already claimed this classification; `report` gains `report_model.py`, `report_summary.py`, `reporter_contract_blocks.py`, `reporter_markdown.py`, `html_report.py`, `junit_report.py`. A larger remaining set (`comparability.py`, `policy_file.py`, `suppression.py`, `surface_graph.py`, `internal_leak.py`, `service_render.py`, `snapshot_io.py`, `build_context.py`, the `dumper_castxml.py`/`dumper_clang_expr.py`/`dumper.py` extract trio, and the `diff_platform.py`/`diff_python.py`/`diff_stdlib_impl.py`/... `compare`-shaped detector family) was investigated and left unclassified: each has a real, already-existing lazy (function-body) import crossing a direction the ADR's stable graph forbids — `compare -> extract` (binary/ELF/PE/Mach-O/Python-API metadata dataclasses still living in their flat parser modules rather than `model/*_facts.py`), `frontends -> policy`/`extract`/`storage` (several `cli_*.py` helpers constructing policy/suppression/build objects directly), `policy -> workflows` (`internal_leak.py`'s helpers, imported lazily from `surface.py`/`post_processing*.py`), and `workflows -> report`/`frontends` (`service.py` re-exporting `service_render.py`, `bundle_side_input.py` reaching into `cli_helpers_compare.py`). Two further candidates (`product_baseline.py`, `stack_binding_diff.py`) passed the same zero-new-findings check under `workflows` but were dropped after review: both only clear the gate there because `workflows`' `may_import` set is wide enough to swallow their current (mis-owned) dependencies — `product_baseline.py`'s own `architecture/debt.yaml` entry already records its real target as `storage`, and `stack_binding_diff.py` is a `compare`-shaped detector (`diff_runtime_bindings` matches baseline/candidate and emits raw `Change` objects) coupled to `binder.py`/`resolver.py`, which are `workflows`-owned shapes `compare` may not import. Classifying either under `workflows` would have satisfied the checker while masking exactly the coupling that needs to move first — left unclassified, consistent with every other module in that "real coupling crosses a forbidden direction" bucket above. No physical file moves, no runtime behavior change — `legacy_paths` entries only.
+- **13 more flat root modules given their ADR-061 responsibility-package classification** — `architecture/modules.yaml`'s `legacy_paths` previously left 57 flat root modules with no `may_import`-direction enforcement at all (a module not physically inside its target package *and* absent from every layer's `legacy_paths` is invisible to `check_architecture.py`'s dependency-direction check, per `_source_layer_for`). This closes a verified-safe subset of that gap, greedily selected so the resulting graph introduces zero new `dependency-direction`/`dependency-cycle`/`unclassified-import` findings (checked against the real repo-wide import graph, not a manual read): `model` gains `qualified_name_segments.py` (a dependency-free leaf shared by `compare`'s `diff_namespaces.py` and `extract`'s `dumper.py`-family callers, reclassified out of `compare` since `extract` may not import `compare`); `extract` gains `dumper_manifest.py`; `policy` gains `pattern_verdicts.py` and the three root-level back-compat shims (`contract_coverage_exit.py`, `exit_decision.py`, `severity.py`) whose own docstrings already claimed this classification; `report` gains `report_model.py`, `report_summary.py`, `reporter_contract_blocks.py`, `reporter_markdown.py`, `html_report.py`, `junit_report.py`. A larger remaining set (`comparability.py` and its `comparability_fields.py`/`comparability_json.py`/`comparability_language_mode.py`/`comparability_sequences.py` helper family, `policy_file.py`, `suppression.py`, `surface_graph.py`, `internal_leak.py`, `service_render.py`, `snapshot_io.py`, `build_context.py`, `stable_abi.py`, the `dumper_castxml.py`/`dumper_clang_expr.py`/`dumper.py` extract trio, and the `diff_platform.py`/`diff_python.py`/`diff_stdlib_impl.py`/... `compare`-shaped detector family) was investigated and left unclassified: each has a real, already-existing lazy (function-body) import crossing a direction the ADR's stable graph forbids — `compare -> extract` (binary/ELF/PE/Mach-O/Python-API metadata dataclasses still living in their flat parser modules rather than `model/*_facts.py`), `frontends -> policy`/`extract`/`storage` (several `cli_*.py` helpers constructing policy/suppression/build objects directly), `policy -> workflows` (`internal_leak.py`'s helpers, imported lazily from `surface.py`/`post_processing*.py`), and `workflows -> report`/`frontends` (`service.py` re-exporting `service_render.py`, `bundle_side_input.py` reaching into `cli_helpers_compare.py`) — plus two modules that are each genuinely mixed rather than owned by one layer: `stable_abi.py` combines a shared data table with real classification/frontend-parsing algorithms (`classify()`, `parse_abi3_version()`), and `comparability.py`'s own helper family straddles `compare` (field-matching/carve-out logic) and `policy` (the gate itself), so both stay unclassified until they're split into their actual owners rather than labeled by their larger, mixed module's convenience (Codex review, PR #942). Two further candidates (`product_baseline.py`, `stack_binding_diff.py`) passed the same zero-new-findings check under `workflows` but were dropped after review: both only clear the gate there because `workflows`' `may_import` set is wide enough to swallow their current (mis-owned) dependencies — `product_baseline.py`'s own `architecture/debt.yaml` entry already records its real target as `storage`, and `stack_binding_diff.py` is a `compare`-shaped detector (`diff_runtime_bindings` matches baseline/candidate and emits raw `Change` objects) coupled to `binder.py`/`resolver.py`, which are `workflows`-owned shapes `compare` may not import. Classifying either under `workflows` would have satisfied the checker while masking exactly the coupling that needs to move first — left unclassified, consistent with every other module in that "real coupling crosses a forbidden direction" bucket above. No physical file moves, no runtime behavior change — `legacy_paths` entries only.
