+++ changelog.d/1788017782_claude_adr061_binary_naming_split.md
@@ -1,3 +1,3 @@
 ### Changed
 
-- **`strip_vendor_hash` split into a new `model`-classified leaf module (`binary_naming.py`), unblocking 7 `compare`-layer detectors under ADR-061** — `binary_utils.py` is a whole-file `extract`-classified module, but `strip_vendor_hash` (and its `_VENDOR_HASH_RE`) is a pure string transform with no I/O and no other dependency, needed by several `compare`-shaped detectors (`diff_platform.py`, `diff_platform_elf_dynamic.py`, `diff_versioning.py`, `diff_wheel_deployment.py`) to pair a vendored library across a rebuild — exactly the shape `qualified_name_segments.py` already established this pattern for. `binary_utils.py` re-exports the name by value for every existing call site; `_canonical_library_key` (which *does* need file I/O, via `_pe_is_dll_content`) imports it back (an allowed `extract -> model` edge). Combined with rewriting several `diff_*.py` modules' lazy/`TYPE_CHECKING` imports of ELF/DWARF/PE/Mach-O/Python metadata dataclasses to their real canonical home (`model.elf_facts`/`model.dwarf_facts`/`model.pe_facts`/`model.macho_facts`/`model.python_facts`/`model.build_mode_facts`) instead of the flat parser-module re-export shims they were reading through (`elf_metadata.py`/`dwarf_metadata.py`/`pe_metadata.py`/`macho_metadata.py`/`python_ext.py`/`python_api.py`/`build_mode.py` — each of those dataclasses' real definition already lived in `model/*_facts.py`, per ADR-061 Phase 5; only the *import path* was pointing at the flat re-export), this closes the `compare -> extract` dependency-direction gap for 7 of the 9 previously-blocked `diff_*.py` detector modules: `diff_numpy_capi.py`, `diff_platform.py`, `diff_platform_elf_dynamic.py`, `diff_python.py`, `diff_python_api.py`, `diff_versioning.py`, `diff_wheel_deployment.py` now classify cleanly as `compare` in `architecture/modules.yaml`. `diff_cpp_patterns.py` stays unclassified — its `_read_elf_soname` genuinely re-parses a binary file via `elf_metadata.parse_elf_metadata` (a real function, not a data shape), which is actual `extract`-owned work happening inside a `compare` module, not an import-path artifact; `diff_stdlib_impl.py` stays unclassified too, for the same reason (`build_mode.build_mode_from_signals`, `surface.compute_public_surface`/`_type_identifiers` are real `extract`/`policy` function calls). No behavior change — `strip_vendor_hash` is re-exported by value (identical function object, pinned by `tests/test_binary_naming.py`), and every fact-type import now resolves to the exact same class it always did.
+- **`strip_vendor_hash` split into a new leaf module physically under `model/` (`model/binary_naming.py`), unblocking 7 `compare`-layer detectors under ADR-061** — `binary_utils.py` is a whole-file `extract`-classified module, but `strip_vendor_hash` (and its `_VENDOR_HASH_RE`) is a pure string transform with no I/O and no other dependency, needed by several `compare`-shaped detectors (`diff_platform.py`, `diff_platform_elf_dynamic.py`, `diff_versioning.py`, `diff_wheel_deployment.py`) to pair a vendored library across a rebuild — the same underlying shape `qualified_name_segments.py` has, but placed directly under `model/` rather than as a flat `legacy_paths` entry, since it's new code with no pre-existing root import path needing a compatibility facade (Codex review, PR #942). `binary_utils.py` re-exports the name by value for every existing call site; `_canonical_library_key` (which *does* need file I/O, via `_pe_is_dll_content`) imports it back (an allowed `extract -> model` edge). Combined with rewriting several `diff_*.py` modules' lazy/`TYPE_CHECKING` imports of ELF/DWARF/PE/Mach-O/Python metadata dataclasses to their real canonical home (`model.elf_facts`/`model.dwarf_facts`/`model.pe_facts`/`model.macho_facts`/`model.python_facts`/`model.build_mode_facts`) instead of the flat parser-module re-export shims they were reading through (`elf_metadata.py`/`dwarf_metadata.py`/`pe_metadata.py`/`macho_metadata.py`/`python_ext.py`/`python_api.py`/`build_mode.py` — each of those dataclasses' real definition already lived in `model/*_facts.py`, per ADR-061 Phase 5; only the *import path* was pointing at the flat re-export), this closes the `compare -> extract` dependency-direction gap for 7 of the 9 previously-blocked `diff_*.py` detector modules: `diff_numpy_capi.py`, `diff_platform.py`, `diff_platform_elf_dynamic.py`, `diff_python.py`, `diff_python_api.py`, `diff_versioning.py`, `diff_wheel_deployment.py` now classify cleanly as `compare` in `architecture/modules.yaml`. `diff_cpp_patterns.py` stays unclassified — its `_read_elf_soname` genuinely re-parses a binary file via `elf_metadata.parse_elf_metadata` (a real function, not a data shape), which is actual `extract`-owned work happening inside a `compare` module, not an import-path artifact; `diff_stdlib_impl.py` stays unclassified too, for the same reason (`build_mode.build_mode_from_signals`, `surface.compute_public_surface`/`_type_identifiers` are real `extract`/`policy` function calls). No behavior change — `strip_vendor_hash` is re-exported by value (identical function object, pinned by `tests/test_binary_naming.py`), and every fact-type import now resolves to the exact same class it always did.
